refactor(coaching-session): readability + DRY cleanups from review
Post-review polish across the title/topics/views work (no behavior change):
- normalize_title_in_update_map: collapse the nested-Option sentinel into one
  if-let with bool::then
- coaching_session_topic::update: use topic.into() + targeted Set like its siblings
  instead of enumerating every ActiveModel field
- extractors: factor the repeated Path-parse block and the 404 literal into shared
  not_found() / parse_path_id_from_parts() helpers
- params: de-dup the clearable-title update-map insert into insert_title_update()
- move the topic-access extractor tests to a separate _tests.rs file (frozen-test
  convention)
- title controller test: result.expect(...) instead of if-let-panic
